dont try to delete a file that doesnt exist silly

+++ recipes/crons.rb
@@ -34,7 +34,7 @@ ruby_block 'create cron jobs' do
 
     # Delete the cron file first, to be sure we don't have old
     # RDS instances in there.
-    File.delete('/var/spool/cron/castor')
+    File.delete('/var/spool/cron/castor') if File.exist?('/var/spool/cron/castor')
     data = JSON.parse(`sudo su castor -c 'aws rds describe-db-instances'`) # ~FC048
     instances = []
     data['DBInstances'].each { |d| instances << d['DBInstanceIdentifier'] }
